C11-159: extract surface.* socket handlers
21 main-switch cases + 28 methods (incl the 4 off-main nonisolated worker
methods v2SurfaceSendText/SendKey/ClearHistory/ReadText, which keep their
nonisolated tier, and 3 surface-only resolve helpers). v2SurfaceTypeDenial is
shared and stays in TerminalController. TerminalController.swift: 15589 -> 14145 LOC.
